Add files via upload
City of Los Angeles Crime Dataset (2020-Present)

Introduction
Welcome to the City of Los Angeles Crime Dataset! This dataset contains information about incidents of crime that occurred in the City of Los Angeles from 2020 to the present. The data has been transcribed from original crime reports that were typed on paper, and it's important to note that this manual transcription process may have resulted in some inaccuracies.

Dataset Details
Dataset Period: 2020 to Present
Last Metadata Update: April 22, 2023

Data Source
The data in this dataset is transcribed from original crime reports. 
LINK - https://catalog.data.gov/dataset/crime-data-from-2020-to-present

Splunk Dashboard
We have created a comprehensive dashboard using Splunk Enterprise to visualize and analyze the City of Los Angeles Crime Dataset. The dashboard provides an interactive and user-friendly interface for exploring various aspects of the dataset. Below is a screenshot of the dashboard:

Splunk Dashboard Screenshot - in Word (2 Dashboard)
1. Crime Analysis for Los Angeles in 2020
2. Victim Analysis for Los Angeles in 2020

Important Notes
This dataset is as accurate as the data in the original crime reports.
Inaccuracies may be present due to the manual transcription process.
Dataset Exploration
Explore the dataset using the Splunk dashboard to gain insights into crime trends, patterns, and other relevant information.